+++ tests/functional/graphing/00-boundaries.t
@@ -15,8 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #-------------------------------------------------------------------------------
-# Test "[visualization]number of cycles" for a workflow with two different cycling
-# intervals.
+# Test graphing for a workflow with two different cycling intervals.
 . "$(dirname "$0")/test_header"
 #-------------------------------------------------------------------------------
 set_test_number 3