+++ tests/functional/validate/28-9999_rollover.t
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 # THIS FILE IS PART OF THE CYLC WORKFLOW ENGINE.
 # Copyright (C) NIWA & British Crown (Met Office) & Contributors.
-# 
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
@@ -15,18 +15,13 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #-------------------------------------------------------------------------------
-# Test fail validation of bad vis node attributes.
+# Test intercycle dependencies.
 . "$(dirname "$0")/test_header"
 #-------------------------------------------------------------------------------
-set_test_number 2
+set_test_number 1
 #-------------------------------------------------------------------------------
-install_workflow "${TEST_NAME_BASE}" "${TEST_NAME_BASE}"
+install_workflow "${TEST_NAME_BASE}" 9999_rollover
 #-------------------------------------------------------------------------------
-TEST_NAME=${TEST_NAME_BASE}
-run_fail "${TEST_NAME}" cylc validate -v "${WORKFLOW_NAME}"
-grep_ok \
-    "Node attributes must be of the form 'key1=value1', 'key2=value2', etc." \
-    "${TEST_NAME}.stderr"
-#-------------------------------------------------------------------------------
+TEST_NAME="${TEST_NAME_BASE}-validate"
+run_fail "${TEST_NAME}" cylc validate --debug "${WORKFLOW_NAME}"
 purge
-exit