+++ CHANGES.md
@@ -83,6 +83,9 @@ that could be held).
 
 ### Fixes
 
+[#4273](https://github.com/cylc/cylc-flow/pull/4273)
+ - Remove obsolete Cylc 7 visualization config section.
+
 [#4272](https://github.com/cylc/cylc-flow/pull/4272) - Workflow visualisation
 data (data-store) now constrained by final cycle point.